+++ src/main/scala/ai/tripl/arc/execute/ConfigExecute.scala
@@ -22,15 +22,14 @@ class ConfigExecute extends PipelineStagePlugin with JupyterCompleter {
 
   val version = Utils.getFrameworkVersion
 
-  val snippet = """{
+  def snippet()(implicit arcContext: ARCContext): String = {
+    s"""{
     |  "type": "ConfigExecute",
     |  "name": "ConfigExecute",
-    |  "environments": [
-    |    "production",
-    |    "test"
-    |  ],
+    |  "environments": [${arcContext.completionEnvironments.map { env => s""""${env}""""}.mkString(", ")}],
     |  "inputURI": "hdfs://*.sql"
     |}""".stripMargin
+  }
 
   val documentationURI = new java.net.URI(s"${baseURI}/execute/#configexecute")