+++ src/main/scala/ai/tripl/arc/extract/StatisticsExtract.scala
@@ -23,16 +23,15 @@ class StatisticsExtract extends PipelineStagePlugin with JupyterCompleter {
 
   val version = Utils.getFrameworkVersion
 
-  val snippet = """{
+  def snippet()(implicit arcContext: ARCContext): String = {
+    s"""{
     |  "type": "StatisticsExtract",
     |  "name": "StatisticsExtract",
-    |  "environments": [
-    |    "production",
-    |    "test"
-    |  ],
+    |  "environments": [${arcContext.completionEnvironments.map { env => s""""${env}""""}.mkString(", ")}],
     |  "inputView": "inputView",
     |  "outputView": "outputView"
     |}""".stripMargin
+  }
 
   val documentationURI = new java.net.URI(s"${baseURI}/extract/#statisticsextract")
 
@@ -41,7 +40,7 @@ class StatisticsExtract extends PipelineStagePlugin with JupyterCompleter {
     import ai.tripl.arc.config.ConfigUtils._
     implicit val c = config
 
-    val expectedKeys = "type" :: "id" :: "name" :: "description" :: "environments" :: "inputView" :: "outputView" :: "persist" :: "approximate" :: "histogram" :: "params" :: Nil
+    val expectedKeys = "type" :: "id" :: "name" :: "description" :: "environments" :: "inputView" :: "outputView" :: "persist" :: "approximate" :: "histogram" :: "hllRelativeSD" :: "params" :: Nil
     val id = getOptionalValue[String]("id")
     val name = getValue[String]("name")
     val description = getOptionalValue[String]("description")
@@ -50,11 +49,12 @@ class StatisticsExtract extends PipelineStagePlugin with JupyterCompleter {
     val approximate = getValue[java.lang.Boolean]("approximate", default = Some(true))
     val histogram = getValue[java.lang.Boolean]("histogram", default = Some(false))
     val persist = getValue[java.lang.Boolean]("persist", default = Some(false))
+    val hllRelativeSD = getValue[java.lang.Double]("hllRelativeSD", default = Some(0.05))
     val params = readMap("params", c)
     val invalidKeys = checkValidKeys(c)(expectedKeys)
 
-    (id, name, description, inputView, outputView, persist, approximate, histogram, invalidKeys) match {
-      case (Right(id), Right(name), Right(description), Right(inputView), Right(outputView), Right(persist), Right(approximate), Right(histogram), Right(invalidKeys)) =>
+    (id, name, description, inputView, outputView, persist, approximate, histogram, hllRelativeSD, invalidKeys) match {
+      case (Right(id), Right(name), Right(description), Right(inputView), Right(outputView), Right(persist), Right(approximate), Right(histogram), Right(hllRelativeSD), Right(invalidKeys)) =>
 
         val stage = StatisticsExtractStage(
           plugin=this,
@@ -67,6 +67,7 @@ class StatisticsExtract extends PipelineStagePlugin with JupyterCompleter {
           persist=persist,
           approximate=approximate,
           histogram=histogram,
+          hllRelativeSD=hllRelativeSD,
         )
 
         stage.stageDetail.put("inputView", inputView)
@@ -75,10 +76,11 @@ class StatisticsExtract extends PipelineStagePlugin with JupyterCompleter {
         stage.stageDetail.put("persist", java.lang.Boolean.valueOf(persist))
         stage.stageDetail.put("approximate", java.lang.Boolean.valueOf(approximate))
         stage.stageDetail.put("histogram", java.lang.Boolean.valueOf(histogram))
+        stage.stageDetail.put("hllRelativeSD", java.lang.Double.valueOf(hllRelativeSD))
 
         Right(stage)
       case _ =>
-        val allErrors: Errors = List(id, name, description, inputView, outputView, persist, approximate, histogram, invalidKeys).collect{ case Left(errs) => errs }.flatten
+        val allErrors: Errors = List(id, name, description, inputView, outputView, persist, approximate, histogram, hllRelativeSD, invalidKeys).collect{ case Left(errs) => errs }.flatten
         val stageName = stringOrDefault(name, "unnamed stage")
         val err = StageError(index, stageName, c.origin.lineNumber, allErrors)
         Left(err :: Nil)
@@ -98,6 +100,7 @@ case class StatisticsExtractStage(
     persist: Boolean,
     approximate: Boolean,
     histogram: Boolean,
+    hllRelativeSD: Double,
   ) extends ExtractPipelineStage {
 
   override def execute()(implicit spark: SparkSession, logger: ai.tripl.arc.util.log.logger.Logger, arcContext: ARCContext): Option[DataFrame] = {
@@ -112,7 +115,7 @@ object StatisticsExtractStage {
 
     val df = spark.table(stage.inputView)
 
-    val statisticsDF = StatisticsUtils.createStatisticsDataframe(df, stage.approximate, stage.histogram)
+    val statisticsDF = StatisticsUtils.createStatisticsDataframe(df, stage.approximate, stage.histogram, stage.hllRelativeSD)
 
     if (arcContext.immutableViews) statisticsDF.createTempView(stage.outputView) else statisticsDF.createOrReplaceTempView(stage.outputView)