+++ src/main/scala/ai/tripl/arc/transform/TensorFlowServingTransform.scala
@@ -34,18 +34,17 @@ class TensorFlowServingTransform extends PipelineStagePlugin with JupyterComplet
 
   val version = Utils.getFrameworkVersion
 
-  val snippet = """{
+  def snippet()(implicit arcContext: ARCContext): String = {
+    s"""{
     |  "type": "TensorFlowServingTransform",
     |  "name": "TensorFlowServingTransform",
-    |  "environments": [
-    |    "production",
-    |    "test"
-    |  ],
+    |  "environments": [${arcContext.completionEnvironments.map { env => s""""${env}""""}.mkString(", ")}],
     |  "inputView": "inputView",
     |  "uri": "http://",
     |  "batchSize": 100,
     |  "outputView": "outputView"
     |}""".stripMargin
+  }
 
   val documentationURI = new java.net.URI(s"${baseURI}/transform/#tensorflowservingtransform")